v1.0.7

+++ FanControl/Util/SMBus.cs
@@ -6,6 +6,7 @@ using System.Runtime.ExceptionServices;
 using System.Runtime.InteropServices;
 using System.Security;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace FanControl
@@ -14,14 +15,29 @@ namespace FanControl
     {
         private static bool sIsOpen = false;
         public static bool IsOpen { 
-            get { return sIsOpen; }
-            set { sIsOpen = value; }
+            get {
+                Monitor.Enter(sLock);
+                bool isOpen = sIsOpen;
+                Monitor.Exit(sLock);
+                return isOpen;
+            }
+            set {
+                Monitor.Enter(sLock);
+                sIsOpen = value;
+                Monitor.Exit(sLock);
+            }
         }
+
+        private static object sLock = new object();
 
         public static bool open()
         {
+            Monitor.Enter(sLock);
             if (sIsOpen == true)
+            {
+                Monitor.Exit(sLock);
                 return true;
+            }               
 
             try
             {
@@ -30,14 +46,19 @@ namespace FanControl
             catch
             {
                 sIsOpen = false;
-            }            
+            }
+            Monitor.Exit(sLock);
             return sIsOpen;
         }
 
         public static void close()
         {
+            Monitor.Enter(sLock);
             if (sIsOpen == false)
+            {
+                Monitor.Exit(sLock);
                 return;
+            }
 
             try
             {
@@ -45,36 +66,62 @@ namespace FanControl
             }
             catch { }
             sIsOpen = false;
+            Monitor.Exit(sLock);
         }
 
         public static byte[] i2cDetect()
         {
+            Monitor.Enter(sLock);
+            if (sIsOpen == false)
+            {
+                Monitor.Exit(sLock);
+                return null;
+            }
+
             try
             {
                 var pByteArrayData = SMBus.getI2CDetect();
                 var retData = SMBus.getBytes(pByteArrayData);
                 SMBus.deleteByteArrayData(pByteArrayData);
+                Monitor.Exit(sLock);
                 return retData;
             }
             catch { }
+            Monitor.Exit(sLock);
             return null;
         }
 
         public static byte[] i2cByteData(byte address, int length)
         {
+            Monitor.Enter(sLock);
+            if (sIsOpen == false)
+            {
+                Monitor.Exit(sLock);
+                return null;
+            }
+
             try
             {
                 var pByteArrayData = SMBus.getI2CByteData(address, length);
                 var retData = SMBus.getBytes(pByteArrayData);
                 SMBus.deleteByteArrayData(pByteArrayData);
+                Monitor.Exit(sLock);
                 return retData;
             }
             catch { }
+            Monitor.Exit(sLock);
             return null;
         }
 
         public static ushort[] i2cWordData(byte address, int length)
         {
+            Monitor.Enter(sLock);
+            if (sIsOpen == false)
+            {
+                Monitor.Exit(sLock);
+                return null;
+            }
+
             try
             {
                 var pByteArrayData = SMBus.getI2CWordData(address, length);
@@ -86,9 +133,11 @@ namespace FanControl
                 {
                     retData[i] = BitConverter.ToUInt16(datas, i * 2);
                 }
+                Monitor.Exit(sLock);
                 return retData;
             }
             catch { }
+            Monitor.Exit(sLock);
             return null;
         }
